GenericChunkedArray now relies on a simple (contiguous ;) std::vector in 64 bits mode (as we don't need 'chunked' memory on 64 bits architectures - but we still keep the name for compatibility with the 32 bits version!)

+++ libs/qCC_db/ccGenericMesh.cpp
@@ -343,7 +343,7 @@ void ccGenericMesh::drawMeOnly(CC_DRAW_CONTEXT& context)
 
 			//we can scan and process each chunk separately in an optimized way
 			//we mimic the way ccMesh beahves by using virtual chunks!
-			unsigned chunks = static_cast<unsigned>(ceil((double)displayedTriNum/(double)MAX_NUMBER_OF_ELEMENTS_PER_CHUNK));
+			unsigned chunks = static_cast<unsigned>(ceil(static_cast<double>(displayedTriNum)/MAX_NUMBER_OF_ELEMENTS_PER_CHUNK));
 			unsigned chunkStart = 0;
 			const colorType* col = 0;
 			for (unsigned k=0; k<chunks; ++k, chunkStart += MAX_NUMBER_OF_ELEMENTS_PER_CHUNK)
